Move contract preview module to the target shape; test result mapping with a Tower mock
Mirrors the same fix applied to the sibling subgraph-preview
implementation:

- Module shape: fold types.rs and the GraphQLQuery derives into mod.rs,
  leaving service.rs with only the tower::Service wrappers, per
  CLAUDE.md's target shape for new rover-client operations.
- The four result_* tests exercising ContractPreviewResult's pure
  field mapping now use a Tower mock (rover_tower::mock_service!)
  instead of spinning up a real MockServer, since they were never
  testing wire-level HTTP behavior.
